Headerbar fix + refactor
