+++ frontend/src/components/headerbar.tsx
@@ -1,23 +1,26 @@
 import styled from "styled-components";
-import { headerBarHeight, sidebarWidth, transitionDuration } from "../config";
-import { ISidebarAwareProps } from "./dashboard";
+import { sidebarWidth, transitionDuration } from "../config";
+import { ISidebarAwareProps } from "./page-wrapper";
 
 /**
  * Header bar for all views. Usually shows control to show/hide sidebar.
  */
+
+// height: ${ headerBarHeight };
+
 export const HeaderBar = styled.div<ISidebarAwareProps>`
   position: fixed;
   display: flex;
 
   top: 0px;
-  left: 0px;
-  width: 100%;
-  height: ${ headerBarHeight };
+  right: 0px;
+
+  height: 50px;
 
   z-index: 1;
 
-  background-color: white;
-  box-shadow: 0 4px 2px -2px gray;
+  background-color: purple;
+  box-shadow: 0 4px 2px -2px rgba(0, 0, 0, 0.1);
 
   flex-direction: row;
   align-items: center;
