+++ frontend/src/components/statistics.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+import { Heading } from "./headings";
+
+/**
+ * A statistics page. Should show things like number of attendees, T-Shirt sizes etc.
+ */
+export const Statistics = () => {
+
+  return (
+    <Heading>Statistics</Heading>
+  );
+
+};
