+++ frontend/src/components/users.tsx
@@ -1,0 +1,15 @@
+import * as React from "react";
+import { Heading } from "./headings";
+
+/**
+ * A page to list all of the users that have registered.
+ */
+export const Users = () => {
+
+  return (
+    <>
+      <Heading>Users</Heading>
+    </>
+  );
+
+};
