+++ frontend/src/index.html
@@ -1,6 +1,7 @@
 <!doctype html>
 <html>
   <head>
+    <link rel="stylesheet" href="node_modules/react-table/react-table.css">
     <meta charset="utf-8">
     <title>tilt</title>
     <style type="text/css">
